refactor: split forum admin tasks
