+++ handlers/forum/remake_thread_stats_task.go
@@ -1,0 +1,8 @@
+package forum
+
+import "github.com/arran4/goa4web/internal/tasks"
+
+// RemakeThreadStatsTask refreshes forum thread statistics.
+type RemakeThreadStatsTask struct{ tasks.TaskString }
+
+var remakeThreadStatsTask = &RemakeThreadStatsTask{TaskString: TaskRemakeStatisticInformationOnForumthread}
